Crs 2179 timeline service refactor (#922)
* CRS-2179 Timeline refactor

* CRS-2179 Removing sds40 calculation option.

* CRS-2179 Removing uneeded comment.

* CRS-2179 Review comments.

* CRS-2179 Fixing maps being modified after calculation.

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/calculatereleasedatesapi/model/external/UnusedDeductionCalculationResponse.kt
@@ -3,6 +3,6 @@ package uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model.external
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.validation.ValidationMessage
 
 data class UnusedDeductionCalculationResponse(
-  val unusedDeductions: Int?,
+  val unusedDeductions: Long?,
   val validationMessages: List<ValidationMessage> = emptyList(),
 )
